Big refactor: all the folder-based tabs support navigation
Needs some tweaking for things without many pages

Big factor as the aircraft/mission/theater tabs were getting a lot of duplicated code, and supporting navigation would have been even more - so now:

* `Tab` is a dedicated interface class with just the base functionality
* There are various other interface classes for optional functionality
* `TabWithDelegate` consistently handles delegation - including when `TabWithNavigation` is implemented.

`TabWithDelegate` is using conditional virtual inheritance, so I expect to use my C++ developer license soon.

+++ src/lib/WintabTablet.cpp
@@ -196,7 +196,7 @@ bool WintabTablet::ProcessMessage(UINT message, WPARAM wParam, LPARAM lParam) {
     p->mState.x = packet.pkX;
     p->mState.y = packet.pkY;
     p->mState.pressure = packet.pkNormalPressure;
-    p->mState.penButtons = packet.pkButtons;
+    p->mState.penButtons = static_cast<uint16_t>(packet.pkButtons);
     return true;
   }
